Fix supply requires
