+++ supply/lib/supply/client.rb
@@ -5,6 +5,9 @@ CredentialsLoader = Google::Auth::CredentialsLoader
 
 require 'net/http'
 
+# This doesn't work properly because of the constants above
+# rubocop:disable Require/MissingRequireStatement
+
 module Supply
   class Client
     # Connecting with Google
