+++ supply/lib/supply/options.rb
@@ -1,5 +1,6 @@
 require 'fastlane_core/configuration/config_item'
 require 'credentials_manager/appfile_config'
+require_relative 'module'
 
 module Supply
   class Options
