+++ supply/lib/supply/uploader.rb
@@ -1,3 +1,8 @@
+require_relative 'module'
+require_relative 'client' 
+require_relative 'apk_listing' 
+require 'fastlane_core/print_table'
+
 module Supply
   class Uploader
     def perform_upload
